// front/src/pages/oee/bug/data.d.ts
export interface TableListItem {
  id?: number;
  reason: string;
  start_time: Date;
  end_time: Date;
}

export interface TablePutItem {
  id?: number;
  reason?: string;
  start_time?: Date;
  end_time?: Date;
}

export interface TableListPagination {
  total: number;
  pageSize: number;
  current: number;
}

export interface TableListData {
  list: TableListItem[];
  pagination: Partial<TableListPagination>;
}

export interface TableListParams {
  status?: string;
  name?: string;
  desc?: string;
  key?: number;
  pageSize?: number;
  currentPage?: number;
  filter?: { [key: string]: any[] };
  sorter?: { [key: string]: any };
}